initial merge totp
