+++ authentik/stages/authenticator_totp/api.py
@@ -9,8 +9,7 @@ from rest_framework.viewsets import GenericViewSet, ModelViewSet
 from authentik.api.authorization import OwnerFilter, OwnerPermissions
 from authentik.core.api.used_by import UsedByMixin
 from authentik.flows.api.stages import StageSerializer
-from authentik.stages.authenticator.plugins.otp_totp.models import TOTPDevice
-from authentik.stages.authenticator_totp.models import AuthenticatorTOTPStage
+from authentik.stages.authenticator_totp.models import AuthenticatorTOTPStage, TOTPDevice
 
 
 class AuthenticatorTOTPStageSerializer(StageSerializer):
