+++ authentik/stages/authenticator_totp/stage.py
@@ -14,8 +14,7 @@ from authentik.flows.challenge import (
     WithUserInfoChallenge,
 )
 from authentik.flows.stage import ChallengeStageView
-from authentik.stages.authenticator.plugins.otp_totp.models import TOTPDevice
-from authentik.stages.authenticator_totp.models import AuthenticatorTOTPStage
+from authentik.stages.authenticator_totp.models import AuthenticatorTOTPStage, TOTPDevice
 from authentik.stages.authenticator_totp.settings import OTP_TOTP_ISSUER
 
 SESSION_TOTP_DEVICE = "totp_device"
